+++ app/ItemImage.php
@@ -11,10 +11,6 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
 
 class ItemImage extends Model
 {
-    const IIP_DEEPZOOM_URL_PREFIX = '/fcgi-bin/iipsrv.fcgi?DeepZoom=';
-    const IIP_DEEPZOOM_URL_SUFFIX = '.dzi';
-    const IIP_IIIF_URL_PREFIX = '/fcgi-bin/iipsrv.fcgi?IIIF=';
-
     protected $fillable = [
         'title',
         'iipimg_url',
@@ -22,22 +18,25 @@ class ItemImage extends Model
         'order',
     ];
 
-    public function getIipimgUrl() {
-        return $this->iipimg_url;
-    }
-
     public function item() {
         return $this->belongsTo(Item::class);
     }
 
-    public function getFullIIPImgURL()
-    {
-        return self::IIP_DEEPZOOM_URL_PREFIX.$this->iipimg_url.self::IIP_DEEPZOOM_URL_SUFFIX;
+    public function getDeepZoomUrl() {
+        return sprintf(
+            '%s/zoom/?path=%s.dzi',
+            config('app.iip_public'),
+            urlencode($this->iipimg_url)
+        );
     }
 
-    public function getFullIIIFImgURL($region = 'full', $size = '!800,800', $rotation = 0, $quality = 'default')
-    {
-        return self::IIP_IIIF_URL_PREFIX.$this->iipimg_url.'/'.$region.'/'.$size.'/'.$rotation.'/'.$quality.'.jpg';
+    public function getPreviewUrl($maxSize = 800) {
+        return sprintf(
+            '%s/preview/?path=%s&size=%d',
+            config('app.iip_public'),
+            urlencode($this->iipimg_url),
+            $maxSize
+        );
     }
 
     public function save(array $options = []) {
@@ -49,11 +48,11 @@ class ItemImage extends Model
         return parent::save($options);
     }
 
-    public function isZoomable() {
-        return $this->iipimg_url !== null;
+    public function getIipimgUrl() {
+        return $this->iipimg_url;
     }
 
     public static function loadValidatorMetadata(ClassMetadata $metadata) {
-        // $metadata->addGetterMethodConstraint('iipimg_url', 'getIipimgUrl', new NotBlank());
+         $metadata->addGetterMethodConstraint('iipimg_url', 'getIipimgUrl', new NotBlank());
     }
 }